Moving QueryIterableResult class into it's own file.
